Move MinifyingSugg into manual_memcpy
